+++ tests/testit/test-base64.R
@@ -14,6 +14,19 @@ assert('base64_encode() encodes the string correctly', {
   (sapply(255:246, function(i) base64_encode(as.raw(255:i))) == ref)
 })
 
+assert('base64_decode() decodes the string correctly', {
+  (sapply(c(1:10, 255:246), function(i) {
+    input = as.raw(1:i)
+    output = base64_encode(input)
+    input2 = base64_decode(output)
+    input %==% input2
+  }))
+})
+
+assert('base64_decode() will not make R crash if the input is not valid', {
+  has_error(base64_decode("lz..??"))
+})
+
 assert('base64_encode_r() returns the same result as base64_encode()', {
   f = R_logo()
   base64_encode_r(f) %==% base64_encode(f)